A harder version of Simon for those without Sight
A game similar to the child’s game Simon, where the player must
remember the pattern of colors shown by the game and then put them back
into the game to receive the next part of the pattern. Instead of
colors, we used sound, and instead of being shown the pattern, the
player must figure out the pattern. The player uses the arrow keys to
guess the “right” direction and then must remember the correct sequence
of directions, all of which are indicated by a sound.